backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.36
# Using pg8000 - pure Python PostgreSQL driver (no C compilation needed)
pg8000==1.31.2
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
alembic==1.13.1
httpx==0.26.0
authlib==1.3.0
itsdangerous==2.1.2
